stepic_web_project

+++ init.sh
@@ -1,11 +1,6 @@
-sudo rm /etc/nginx/sites-enabled/default
-sudo ln -s /home/box/web/etc/nginx.conf /etc/nginx/sites-enabled/default
+if [ -e /etc/nginx/sites-enabled/default ] ; then
+    sudo rm /etc/nginx/sites-enabled/default
+fi
+
+sudo ln -sf /home/box/web/etc/nginx.conf /etc/nginx/sites-enabled/test.conf
 sudo /etc/init.d/nginx restart
-#sudo ln -s /home/box/web/etc/gunicorn.conf /etc/gunicorn.d/gunicorn.conf
-#sudo ln -s /home/box/web/etc/django.conf /etc/gunicorn.d/django.conf
-#sudo /etc/init.d/gunicorn restart
-sudo /etc/init.d/mysql restart
-mysql -uroot -e "CREATE DATABASE stepic_qa;"
-mysql -uroot -e "CREATE USER 'box'@'localhost' IDENTIFIED BY '1234';"
-mysql -uroot -e "GRANT ALL PRIVILEGES ON stepic_qa.* TO 'box'@'localhost';"
-mysql -uroot -e "FLUSH PRIVILEGES;"
